+++ docs/app/layout.tsx
@@ -1,12 +1,13 @@
 import { getEnhancedPageMap } from '@components/get-page-map'
 import { NextraLogo, VercelLogo } from '@components/icons'
+import { ChatButton } from '@components/inkeep-chat-button'
 import cn from 'clsx'
 import type { Metadata } from 'next'
 import NextImage from 'next/image'
 import { Footer, Layout, Link, Navbar } from 'nextra-theme-docs'
 import { Anchor, Banner, Head } from 'nextra/components'
-import type { FC, ReactNode } from 'react'
-import xyflow from './showcase/_logos/xyflow.png'
+import type { FC } from 'react'
+import inkeep from './showcase/_logos/inkeep.png'
 import './globals.css'
 
 export const metadata: Metadata = {
@@ -50,12 +51,11 @@ export const metadata: Metadata = {
 }
 
 const banner = (
-  <Banner dismissible={false}>
-    🎉 Nextra 4.0 is released. dimaMachina is looking{' '}
-    <Link href="https://github.com/dimaMachina" className="text-current!">
-      for a new job or consulting
+  <Banner>
+    🎉 Nextra 4.0 has been released.{' '}
+    <Link href="https://the-guild.dev/blog/nextra-4" className="text-current!">
+      Read blogpost
     </Link>
-    .
   </Banner>
 )
 const navbar = (
@@ -91,14 +91,13 @@ const footer = (
   </Footer>
 )
 
-const RootLayout: FC<{
-  children: ReactNode
-}> = async ({ children }) => {
+const RootLayout: FC<LayoutProps<'/'>> = async ({ children }) => {
   const pageMap = await getEnhancedPageMap()
   return (
     <html lang="en" dir="ltr" suppressHydrationWarning>
       <Head />
       <body>
+        <ChatButton />
         <Layout
           banner={banner}
           navbar={navbar}
@@ -111,10 +110,11 @@ const RootLayout: FC<{
             extraContent: (
               <>
                 <b className="mt-2 text-xs">Sponsored by:</b>
-                <Anchor href="https://xyflow.com?utm_source=nextra.site&utm_campaign=nextra&utm_content=sidebarLink">
+                <Anchor href="https://inkeep.com?utm_source=nextra.site&utm_campaign=nextra&utm_content=sidebarLink">
                   <NextImage
-                    src={xyflow}
-                    alt="Wire your ideas with xyflow!"
+                    src={inkeep}
+                    title="AI Agents that get real work done"
+                    alt="AI Agents that get real work done"
                     className="nextra-border rounded-sm border"
                   />
                 </Anchor>