+++ docs/next.config.ts
@@ -80,6 +80,10 @@ const nextConfig = withNextra({
     // ESLint behaves weirdly in this monorepo.
     ignoreDuringBuilds: true
   },
+  env: {
+    NEXT_PUBLIC_INKEEP_API_KEY:
+      'dee399c7f7ac40b9de0d0b85ca32959953b9ff7c9fc8d96c'
+  },
   redirects: async () => [
     {
       source: '/docs/guide/:slug(typescript|latex|tailwind-css|mermaid)',